Make DIFM starting point components reusable
